backend/requirements.txt
Flask==3.0.3
Flask_Cors==4.0.1
mysql_connector_repackaged==0.3.1
python-dotenv==1.0.1
pandas